For the 3D printer
This is the stage for the electronics on the semi-automatic spectra machine setup. This stage will be drilled onto the front of the platform and house the motor driver and the Arduino. For the rest of the circuitry, it will be housed on breadboards that are attached via adhesion to the platform.